something is working2

+++ Task_1_1_3/src/main/java/ru/nsu/fitkulin/Number.java
@@ -2,7 +2,7 @@ package ru.nsu.fitkulin;
 
 public class Number extends Expression {
     double value;
-    public Number(int value) {
+    public Number(double value) {
         super();
         this.value = value;
     }
